update tests for ChatGenerator
